CAL: updated calibration code, working for RX and TX
- FPGA v10.2

+++ libloragw/inc/loragw_cal.h
@@ -35,9 +35,9 @@ License: Revised BSD License, see LICENSE.TXT file include in the project
 struct lgw_sx125x_cal_rx_result_s {
     int8_t amp;
     int8_t phi;
-    uint8_t rej;
-    uint8_t rej_init;
-    uint8_t snr;
+    uint16_t rej;
+    uint16_t rej_init;
+    uint16_t snr;
 };
 
 struct lgw_sx125x_cal_tx_result_s {
@@ -45,8 +45,8 @@ struct lgw_sx125x_cal_tx_result_s {
     uint8_t mix_gain;
     int8_t offset_i;
     int8_t offset_q;
-    uint8_t rej;
-    uint8_t sig;
+    uint16_t rej;
+    uint16_t sig;
 };
 
 /* -------------------------------------------------------------------------- */
